+++ tests/test_huestr.py
@@ -32,14 +32,14 @@ def test_reset_chained():
 
 def test_hues_creation():
   obj = HueString('woot')
-  assert isinstance(obj, str)
+  # assert isinstance(obj, str)
   assert obj.colorized == 'woot'
 
 
 def test_hues_auto_stacking():
   obj = HueString('woot').cyan.bg_green
-  assert isinstance(obj, str)
-  assert obj._HueString__hue_stack == (FG.cyan, BG.green)
+  # assert isinstance(obj, str)
+  assert obj.hues == (FG.cyan, BG.green)
 
 
 def test_hues_dynamic_props_exceptions():